+++ pom.xml
@@ -158,6 +158,9 @@
                 <version>3.2.0</version>
                 <configuration>
                     <detectLinks/>
+                    <links>
+                        <link>https://docs.spring.io/spring-framework/docs/current/javadoc-api/</link>
+                    </links>
                 </configuration>
                 <executions>
                     <execution>
